+++ oshi-core/src/main/java/oshi/hardware/CentralProcessor.java
@@ -148,14 +148,23 @@ public interface CentralProcessor {
      * 3, representing 1, 5, and 15 minutes. The system load average is the sum of
      * the number of runnable entities queued to the available processors and the
      * number of runnable entities running on the available processors averaged over
-     * a period of time. The way in which the load average is calculated is
-     * operating system specific but is typically a damped time-dependent average.
-     * If the load average is not available, a negative value is returned. This
-     * method is designed to provide a hint about the system load and may be queried
-     * frequently.
+     * a period of time.
      * <p>
-     * The load average may be unavailable on some platforms (e.g., Windows) where
-     * it is expensive to implement this method.
+     * This method is designed to provide a hint about the system load and may be
+     * queried frequently.
+     * <p>
+     * The way in which the load average is calculated is operating system specific
+     * but is typically a damped time-dependent average. Linux includes processes
+     * waiting for system resources such as disks, while macOS and Unix consider
+     * only processes waiting for CPU.
+     * <p>
+     * Windows does not provide a load average. Users may set the configuration
+     * property {@code oshi.os.windows.loadaverage} to {@code true} to start a
+     * daemon thread which will provide a similar metric.
+     * <p>
+     * The load average may be unavailable on some platforms (e.g., Windows without
+     * the above configuration). If the load average is not available, a negative
+     * value is returned.
      *
      * @param nelem
      *            Number of elements to return.